Authorization Part1 + Relationship btw User and Player (Part2)

+++ src/main/java/com/mike/waf/service/UserService.java
@@ -2,9 +2,11 @@ package com.mike.waf.service;
 
 import com.mike.waf.model.entities.User;
 import com.mike.waf.repository.UserRepository;
+import com.mike.waf.security.AuthorizationValidatorService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -16,6 +18,7 @@ public class UserService implements IService<User> {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private final AuthorizationValidatorService authorizationValidatorService;
 
     @Override
     public Optional<User> findById(UUID id) {
@@ -38,6 +41,7 @@ public class UserService implements IService<User> {
         if (user.getId() == null) {
             throw new IllegalArgumentException("User must have an id");
         }
+        authorizationValidatorService.validate(user);
         userRepository.save(user);
     }
 
@@ -46,10 +50,17 @@ public class UserService implements IService<User> {
         if (user.getId() == null) {
             throw new IllegalArgumentException("User must have an id");
         }
+        authorizationValidatorService.validate(user);
         userRepository.delete(user);
     }
 
+    @Transactional
     public void deleteByUsername(String username) {
+        var user = userRepository.findByUsername(username);
+        if (user == null) {
+            throw new IllegalArgumentException("User not found");
+        }
+        authorizationValidatorService.validate(user.get());
         userRepository.deleteByUsername(username);
     }
